feat(posts): add SQS notification for post creation
Add a new method `sendPostNotification` in the SQS class to send notifications when a post is created. This method is integrated into the post creation lambda to notify the recipient when a post is made about them. The `@oppfy/sqs` package is added as a dependency to support this functionality.